# backend/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internationalized culture / e-commerce / custom-travel platform for Jingdezhen ceramic art.\nBackend in Go + Fiber; money as BIGINT minor units (CNY base, USD/EUR/GBP presentment).",
        "title": "Jingdezhen Ceramics Platform API",
        "termsOfService": "https://jingdezhen.test/terms",
        "contact": {
            "name": "Jingdezhen Ceramics Platform Team"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/profile/addresses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all shipping addresses for the signed-in user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile",
                    "addresses"
                ],
                "summary": "List the current user's shipping addresses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{data: []models.UserAddress}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a shipping address to the signed-in user's address book.\nSet is_default=true to also make it the default.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile",
                    "addresses"
                ],
                "summary": "Create a shipping address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Address to create (country is ISO 3166-1 alpha-2)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.CreateAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.UserAddress"
                        }
                    },
                    "400": {
                        "description": "Invalid body / validation",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile/addresses/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches one of the signed-in user's addresses by ID. An address\nnot owned by the user returns 404 (no cross-user access).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile",
                    "addresses"
                ],
                "summary": "Get one shipping address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.UserAddress"
                        }
                    },
                    "400": {
                        "description": "Invalid address ID",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Address not found (or not owned by user)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replaces all editable fields of an address. is_default is honoured\nhere too (so a full-replace client can set it in one call), though\nit's also manageable via the set-default endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile",
                    "addresses"
                ],
                "summary": "Replace a shipping address (PUT semantics)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Full replacement of editable fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.UpdateAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.UserAddress"
                        }
                    },
                    "400": {
                        "description": "Invalid address ID / body / validation",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Address not found (or not owned by user)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes one of the signed-in user's addresses.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile",
                    "addresses"
                ],
                "summary": "Delete a shipping address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (empty body)"
                    },
                    "400": {
                        "description": "Invalid address ID",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Address not found (or not owned by user)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile/addresses/{id}/default": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks one of the signed-in user's addresses as the default\n(used as the checkout prefill).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile",
                    "addresses"
                ],
                "summary": "Set the default shipping address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003caccess_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Address ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{message: \\\"Default address updated\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid address ID",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Address not found (or not owned by user)",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/jingdezhen-ceramics-backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jingdezhen-ceramics-backend_internal_models.CreateAddressRequest": {
            "type": "object",
            "required": [
                "city",
                "country",
                "line1",
                "recipient"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 100
                },
                "country": {
                    "description": "alpha-2, e.g. \"US\"",
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "line1": {
                    "type": "string",
                    "maxLength": 255
                },
                "line2": {
                    "type": "string",
                    "maxLength": 255
                },
                "phone": {
                    "type": "string",
                    "maxLength": 30
                },
                "postal_code": {
                    "type": "string",
                    "maxLength": 30
                },
                "recipient": {
                    "type": "string",
                    "maxLength": 255
                },
                "region": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
        "jingdezhen-ceramics-backend_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Optional additional details",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "jingdezhen-ceramics-backend_internal_models.UpdateAddressRequest": {
            "type": "object",
            "required": [
                "city",
                "country",
                "line1",
                "recipient"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 100
                },
                "country": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "line1": {
                    "type": "string",
                    "maxLength": 255
                },
                "line2": {
                    "type": "string",
                    "maxLength": 255
                },
                "phone": {
                    "type": "string",
                    "maxLength": 30
                },
                "postal_code": {
                    "type": "string",
                    "maxLength": 30
                },
                "recipient": {
                    "type": "string",
                    "maxLength": 255
                },
                "region": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
        "jingdezhen-ceramics-backend_internal_models.UserAddress": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "description": "ISO 3166-1 alpha-2",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "line1": {
                    "type": "string"
                },
                "line2": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003caccess_token\u003e\" — JWT access token from POST /auth/login (or OAuth). Required for all /admin/* and customer-authed endpoints; omitted for public /catalog/* reads.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}